Serialization: Cleaner dir&file names.
