+++ TypeSystem/Serialization/json.h
@@ -26,121 +26,16 @@ SOFTWARE.
 #ifndef CURRENT_TYPE_SYSTEM_SERIALIZATION_JSON_H
 #define CURRENT_TYPE_SYSTEM_SERIALIZATION_JSON_H
 
-#include <chrono>  // For the future use of "primitive_types.dsl.h".
+#include "serialization.h"
 
-#include "base.h"
-#include "exceptions.h"
-
-// Current: as human- and JavaScript-readable as possible, with strong struct signature checking.
-// Used internally, and is the default for exposed API endpoints.
-// * `CURRENT_STRUCT`-s as objects with field names as keys.
-// * `map<string, X>` as `{"k1":<v1>, "k2":<v2>, ...}`, `map<K, V>` as `[[<k1>, <v1>], [<k2>, <v2>], ...]`.
-// * `Variant<A, B, C> x = A(...)` as `{"A":<body>, "":<type ID of A>}`.
-//
-// Minimalistic: same as `Current` above except for treating `Variant`s.
-// Less strict. Used primarily to import data into Current from a trusted source, ex. a local Python script.
-// * Does not serialize Type ID into the empty key.
-// * Relies on the name of the top-level object only when doing `Variant` deserialization.
-//
-// NewtonsoftFSharp: the format friendly with `import Newtonsoft.JSON` from F#.
-// Used to communicate with F# and/or Visual Studio.
-// Difference from `Current`:
-// * Variant types are stored as `{"Case":"A", "Fields":[ { <fields of A> } ]}`.
-// * TODO(dkorolev): Confirm `map<K, V>` does what it should. Untested as of now; `vector<T>` works fine.
-
-// Basic types.
-#include "types/primitive.h"
-#include "types/enum.h"
-
-// STL containers.
-#include "types/vector.h"
-#include "types/pair.h"
-#include "types/map.h"
-
-// Current types.
-#include "types/current_typeid.h"
-#include "types/current_struct.h"
-#include "types/optional.h"
-#include "types/variant.h"
-
-namespace current {
-namespace serialization {
-
-namespace json {
-
-template <class J, typename T>
-void ParseJSONViaRapidJSON(const std::string& json, T& destination) {
-  rapidjson::Document document;
-
-  if (document.Parse(json.c_str()).HasParseError()) {
-    throw InvalidJSONException(json);
-  }
-
-  load::LoadFromJSONImpl<T, J>::Load(&document, destination, "");
-}
-
-template <class J = JSONFormat::Current, typename T>
-inline std::string JSON(const T& source) {
-  JSONStringifier<J> json_stringifier;
-  Serialize(json_stringifier, source);
-  return json_stringifier.ResultingJSON();
-}
-
-template <class J = JSONFormat::Current>
-inline std::string JSON(const char* special_case_bare_c_string) {
-  return JSON<J>(std::string(special_case_bare_c_string));
-}
-
-template <typename T, class J = JSONFormat::Current>
-inline void ParseJSON(const std::string& source, T& destination) {
-  try {
-    ParseJSONViaRapidJSON<J>(source, destination);
-    CheckIntegrity(destination);
-  } catch (UninitializedVariant) {
-    throw JSONUninitializedVariantObjectException();
-  }
-}
-
-template <typename T, class J = JSONFormat::Current>
-inline void PatchJSON(T& object, const std::string& json) {
-  try {
-    CheckIntegrity(object);  // TODO(dkorolev): Different exception for "was uninitialized before"?
-    ParseJSONViaRapidJSON<JSONPatcher<J>>(json, object);
-    CheckIntegrity(object);
-  } catch (UninitializedVariant) {
-    throw JSONUninitializedVariantObjectException();
-  }
-}
-
-template <typename T, class J = JSONFormat::Current>
-inline T ParseJSON(const std::string& source) {
-  T result;
-  ParseJSON<T, J>(source, result);
-  return result;
-}
-
-}  // namespace json
-}  // namespace serialization
-
-// Keep top-level symbols both in `current::` and in global namespace.
-using serialization::json::JSON;
-using serialization::json::ParseJSON;
-using serialization::json::PatchJSON;
-using serialization::json::JSONFormat;
-using serialization::json::TypeSystemParseJSONException;
-using serialization::json::JSONSchemaException;
-using serialization::json::InvalidJSONException;
-using serialization::json::JSONUninitializedVariantObjectException;
-
-}  // namespace current
-
-using current::JSON;
-using current::ParseJSON;
-using current::PatchJSON;
-using current::JSONFormat;
-using current::TypeSystemParseJSONException;
-using current::JSONSchemaException;
-using current::InvalidJSONException;
-using current::JSONUninitializedVariantObjectException;
+#include "json/enum.h"
+#include "json/map.h"
+#include "json/optional.h"
+#include "json/pair.h"
+#include "json/primitives.h"
+#include "json/struct.h"
+#include "json/typeid.h"
+#include "json/variant.h"
+#include "json/vector.h"
 
 #endif  // CURRENT_TYPE_SYSTEM_SERIALIZATION_JSON_H
